+++ Network/API/Telegram/Bot/Object/Chat.hs
@@ -1,7 +1,6 @@
 module Network.API.Telegram.Bot.Object.Chat (module Exports, Chat) where
 
 import Network.API.Telegram.Bot.Object.Chat.Group as Exports
-import Network.API.Telegram.Bot.Object.Chat.Conversation as Exports
 import Network.API.Telegram.Bot.Object.Chat.Channel as Exports
 
 data Chat